# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo REST API",
        "description": "Простое CRUD API для задач и тегов",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Список тегов",
                "operationId": "e2dbd17f7f4d6f789398e1204b76f0d7",
                "responses": {
                    "200": {
                        "description": "Список тегов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Создать тег",
                "operationId": "1931bad8ef209e3bbf89f33afad91411",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Тег создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Получить тег по ID",
                "operationId": "f7ef72d520711ee9bfd33534202c53c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тег найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Тег не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Удалить тег",
                "operationId": "7a83373773dc2833b2787028d7867d22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Тег удалён"
                    },
                    "404": {
                        "description": "Тег не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tags"
                ],
                "summary": "Обновить тег",
                "operationId": "1c7b374279f22ce0c9c1358924cb08cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Тег обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Тег не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "required": [
                    "id",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Personal"
                    }
                },
                "type": "object"
            },
...
            "StoreTagRequest": {
                "required": [
                    "description"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateTagRequest": {
                "required": [
                    "description"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "BaseSuccessEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {}
                },
                "type": "object"
            },
            "BaseErrorEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {}
                },
                "type": "object"
            },
...
            "TagSuccessResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseSuccessEnvelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "TagListResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseSuccessEnvelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ErrorResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseErrorEnvelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "string",
                                "example": "Задача с айди 1 не найдена",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation failed. (title field is required)"
                    }
                },
                "type": "object"
            }
        }
    },
...
}